Like type

+++ app/src/main/java/com/dabenxiang/mimi/model/api/vo/MembersPostCommentItem.kt
@@ -19,7 +19,7 @@ data class MembersPostCommentItem(
     @SerializedName("likeCount")
     val likeCount: Long?,
     @SerializedName("likeType")
-    val likeType: Long?,
+    val likeType: Int?,
     @SerializedName("postAvatarAttachmentId")
     val postAvatarAttachmentId: Long?,
     @SerializedName("postName")
